Move DataObject snapshot tests into the snapshot test directory
Snapshot tests go alongside the other snapshot-based tests under
``tests/snapshots/`` and use the shared ``SNAPSHOT_SETTINGS`` +
``run_test_for_falsifying_example`` helpers. The
``test_combinators.py::test_data_draw`` snapshot is also regenerated to
reflect the new ``DataObject(draws=[...])`` rendering.
